SuzdalAI/requirements.txt
fastapi
uvicorn
requests
pandas
python-dotenv
tenacity
langchain
langchain-community
langchain-gigachat
